setting up logic for the appropriate urls and views

+++ Notifications/src/main/java/com/NitsaBedianashvili/Notifications/repository/NotificationRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NotificationRepo extends JpaRepository<NotificationPreference, Integer> {
+public interface NotificationRepo extends JpaRepository<NotificationPreference, Long> {
 
 
 }
